# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/schedule/task": {
            "post": {
                "description": "Create a new task",
                "summary": "Create a new task",
                "parameters": [
                    {
                        "description": "Task",
                        "name": "task",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Task"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.TaskDTO"
                        }
                    }
                }
            }
        },
        "/schedule/task/{id}": {
            "get": {
                "description": "Get a task by ID",
                "summary": "Get a task by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.TaskDTO"
                        }
                    }
                }
            },
            "put": {
                "description": "Update a task",
                "summary": "Updates a task",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Task",
                        "name": "task",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Task"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.TaskDTO"
                        }
                    }
                }
            },
            "post": {
                "description": "CheckIn on a task",
                "summary": "CheckIn on a task",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.CheckInDTO"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a task",
                "summary": "Deletes a task",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.TaskDTO"
                        }
                    }
                }
            }
        },
        "/schedule/{date}": {
            "get": {
                "description": "Get a schedule by date",
                "summary": "Get schedule by date",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Date",
                        "name": "date",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.TaskDTO"
                        }
                    }
                }
            }
        },
        "/schedule/{date}/{groupid}": {
            "get": {
                "description": "Get a list of tasks by date and group",
                "summary": "Get the tasks by date and group",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Date",
                        "name": "date",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Group ID",
                        "name": "groupid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.TaskDTO"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.CheckInDTO": {
            "type": "object",
            "properties": {
                "checkedIn": {
                    "type": "boolean"
                },
                "checkinId": {
                    "type": "integer"
                },
                "checkinTime": {
                    "type": "string"
                },
                "taskId": {
                    "type": "integer"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
        "model.Location": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "lat": {
                    "type": "number"
                },
                "longtitude": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "model.Task": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "endTime": {
                    "type": "string"
                },
                "groupId": {
                    "type": "integer"
                },
                "location": {
                    "description": "has to be object",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Location"
                        }
                    ]
                },
                "name": {
                    "type": "string"
                },
                "startTime": {
                    "type": "string"
                }
            }
        },
        "model.TaskDTO": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "endTime": {
                    "type": "string"
                },
                "groupId": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "location": {
                    "description": "has to be object",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Location"
                        }
                    ]
                },
                "name": {
                    "type": "string"
                },
                "startTime": {
                    "type": "string"
                }
            }
        }
    }
}